HF3D: Add servo override and status commands
This commit extends the servo command with the following options:

    servo status
    servo override
    servo override <value>
    servo override <N> <value>

Status shows the current position of all servos.

Servo override forces the servo position to the given value [ms].
If <N> is given, the value applies to servo #N, otherwise to all
servos. The special keyword 'off' as a value turns override off.

Servo override is functional only when not armed.